Delay calling `build.getExecutor()`.
Due to the implementation in Jenkins core, this call is actually
very expensive (O(N*log(N)) in the number of executors, with
hefty constant factors, due to the implementation of
`Jenkins::getComputers`). While that could be improved, we can
also avoid calling it until we're actually deciding to interrupt
a build, thus saving the overhead.